Refactor modules to separate evaluation and dataset
