+++ aieng-eval-agents/aieng/agent_evals/knowledge_qa/__init__.py
@@ -24,12 +24,7 @@ from aieng.agent_evals.tools import (
 )
 
 from .agent import KnowledgeAgentManager, KnowledgeGroundedAgent
-from .evaluation import (
-    DeepSearchQADataset,
-    DeepSearchQAEvaluator,
-    DSQAExample,
-    EvaluationResult,
-)
+from .evaluation import DeepSearchQAEvaluator, EvaluationResult
 
 
 __all__ = [
