+++ aieng-eval-agents/aieng/agent_evals/knowledge_qa/data/__init__.py
@@ -1,0 +1,13 @@
+"""Data loading and management for knowledge QA evaluation.
+
+This module provides tools for loading and managing benchmark datasets
+like DeepSearchQA.
+"""
+
+from .deepsearchqa import DeepSearchQADataset, DSQAExample
+
+
+__all__ = [
+    "DSQAExample",
+    "DeepSearchQADataset",
+]
