+++ implementations/knowledge_qa/gradio_app.py
@@ -13,10 +13,8 @@ import uuid
 from typing import Any, Generator
 
 import gradio as gr
-from aieng.agent_evals.knowledge_qa import (
-    DeepSearchQADataset,
-    KnowledgeAgentManager,
-)
+from aieng.agent_evals.knowledge_qa import KnowledgeAgentManager
+from aieng.agent_evals.knowledge_qa.data import DeepSearchQADataset
 from dotenv import load_dotenv
 from gradio.components.chatbot import ChatMessage
